requirements.txt
aiohttp==3.7.4.post0
aiohttp-middlewares==1.1.0
aiohttp-retry==2.4.0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
colorama==0.4.4
DataProperty==0.50.1
deprecation==2.1.0
diskcache==5.2.1
dkpro-cassis==0.5.2
dkpro-cassis-tools==0.0.6
dnspython==2.1.0
email-validator==1.1.2
environ-config==21.2.0
html5lib==1.1
idna==3.2
importlib-resources==5.1.4
isodate==0.6.0
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lxml==4.6.3
mbstrdecoder==1.0.1
more-itertools==8.8.0
msgfy==0.1.0
multidict==5.1.0
numpy==1.20.3
openapi-core==0.13.6
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==20.9
pandas==1.2.4
parse==1.19.0
pathvalidate==2.4.1
prettytable==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytablewriter==0.60.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
rororo==2.2.2
sequence-transfer==0.1.0
six==1.16.0
sortedcontainers==2.4.0
tabledata==1.1.4
tcolorpy==0.0.9
toposort==1.6
typepy==1.1.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
web-anno-tsv==0.0.1
webencodings==0.5.1
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.4.1
